refactor(gui): extract DSP-applet + CAT + DAX-IQ wiring from constructor (#3351 Phase 2d)
Tenth PR of the #3351 series — completes the constructor decomposition.
Same pattern throughout: each block becomes a named private wireXxx()
method called at exactly the original constructor position. Three
additive MainWindow.h declarations.

- wireDspApplets() → new MainWindow_DspApplets.cpp (~525 lines):
  PooDoo RX chain status tiles, P/CW applet meters, PHNE, EQ, the
  client DSP applet family (EQ/Comp/Gate/DeEss/Tube/Reverb/AetherDSP/
  PUDU — TX and RX tiles), TX signal-chain applet, PUDU monitor,
  RX chain edit + bypass. Window-side wiring (applets ↔ AudioEngine);
  the TU header says so.
- wireCatPorts() → MainWindow_Session.cpp (~100 lines): the unified
  CAT port slots — session-bound, joins the RadioSession seed file.
- wireDaxIq() → MainWindow_Session.cpp (~75 lines): DAX IQ wiring for
  no-audio-bridge platforms — likewise session-bound.

Per the Phase 2c lesson, all three cut ranges were checked for stacked
earlier-phase wireXxx() calls before extraction (none present) and
ifdef balance (all zero). Includes carried explicitly.

The constructor is now ~1,000 lines: UI construction order, geometry
restore, and small chrome wiring — a legible composition root. The
constructor-decomposition phase of #3351 is complete.

MainWindow.cpp: 8,766 → 8,075 lines (series: 19,474 → 8,075, −58.5%).
Verified: full build clean, all 33 ctest suites pass, a11y lint clean.

Refs #3351 #3445. Principle XI.
